Misc/IncludeMismatch: bug fix
The file name may be build up of several parts. The sniff - until now - would only look at the first text string token for the extension, while the extension will normally be in the _last_ text string token.

This commit changes the sniff to walk back from the end of the statement instead of forward.

Includes unit test.