BadFunctions/Backticks: bug fix - report on each variable
The sniff would only report on the first variable found in the shell command, not on each variable.

Even though there would be a notice, the level could have been `warning` as the first variable seen was a non-user input variable, while a more serious `error` for a subsequently used user input variable would not be reported.

This has now been fixed by changing the check for a variable to a loop which will report a separate error/warning for each variable encountered in the command.