BadFunctions/EasyRFI: efficiency fix
As the tokens to search for don't change during a PHPCS run, it is inefficient to use the "expensive" `array_merge()` function 1) within a `while` loop and 2) every time the sniff is triggered by an include/require token.

The set of tokens to search for can just as easily be set only once before the sniff is ever triggered and doing so will make the sniff faster.