BadFunctions/EasyRFI: bug fix - fix detecting of start/end of the statement [1]
Prevent false negatives when an `include`/`require` statement combines parentheses with concatentation outside parentheses.

Checking whether the next non-empty token is an open parenthesis could cause false negatives as there may be additional paths of the path _after_ the parenthesized part of the statement.

The only reason why this wasn't a problem up to now is because of a bug in the sniff determining `$s`.

The `findNext()` function searches in the token stack and treats the `$start` parameter as _inclusive_.

That means that when searching for the next non-empty token, passing the `$stackPtr` to an include/require statement would _always_ return the `$stackPtr` and not the next non-empty token _after_ the `$stackPtr` which could have been an open parenthesis (or not).

Taking the logic related to the parentheses out of the sniff prevents this issue.