BadFunctions/EasyRFI: bug fix - fix detecting of start/end of the statement [3]
An `include`/`require` statement can be used within template files to include another template and doesn't need a closing semi-colon in that case.

That situation was so far not being considered by the sniff and the sniff could in that case search way to far and report on completely unrelated statements after the include.