Remove `Drupal8` and `Symfony2` support
As discussed in 70, as `Drupal8` and `Symfony2` support hasn't been implemented yet anyway, let's remove the classes inferring that there is support for these frameworks to prevent confusion and simplify maintainance.

Note: this can be considered a breaking change as, in the unlikely situation that someone would have set `CmsFramework` to `Drupal8` or `Symfony2` in their custom ruleset, running PHPCS will now throw a fatal error.
In other words, this commit should go into a major release and will need a changelog entry.