BadFunctions/EasyRFI: use the build-in PHPCS functionality [2]
The PHPCS native `PHP_CodeSniffer\Util\Tokens` class contains a number of useful token groups.

For this particular sniff, the `Tokens::$includeTokens` applies and contains all the relevant tokens.

It is generally a good idea to use the build-in token groups, as when something would change in how PHP and/or PHPCS tokenizes certain constructs, those groups will be updated too.